ingestion/requirements.txt
pdfplumber>=0.9.0
python-docx>=1.1.0
supabase>=2.3.0
python-dotenv>=1.0.0
click>=8.1.0
